requirements.txt
# Core simulation
numpy>=1.24,<2.0
pandas>=2.0,<3.0
pyyaml>=6.0,<7.0
scipy>=1.10,<2.0

# Visualization
matplotlib>=3.7,<4.0
plotly>=5.18,<6.0

# Dashboard
streamlit>=1.28,<2.0

# Development / testing
pytest>=7.4
pytest-cov>=4.1
